opkg_download: add escape_url option for CURL
When enabled, this option will use curl_easy_escape to escape the path
portion of the url. It will only do this for http/https urls.

This is a fix to support hosting feeds with providers similar to AWS S3.
AWS S3 handles all unescaped plus signs [+] in the path as spaces.

AWS will look for file+name as file%20name instead of file%2Bname unless
it is already escaped.